requirements.txt
fastapi
uvicorn
gitpython
requests
python-dotenv
pydantic
aiofiles
tqdm
numpy
scipy
python-Levenshtein
faiss-cpu
sentence-transformers
scikit-learn
chardet
langgraph
gitpython
pandas
openpyxl
python-dotenv
google-generativeai
matplotlib
radon          
chardet
requests
google.genai
supabase
postgrest
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
faker